pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yarayaml"]

[project]
name = "yara-yaml"
version = "0.0.15"
authors = [
    { name="Darren Spruell", email="phatbuckett@gmail.com" },
]
description = "YARA rules from YAML and Jinja templates"
dependencies = [
    "Jinja2 >= 3.1",
    "PyYAML",
    "tabulate",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    # "Development Status :: 4 - Beta",
    # "Development Status :: 5 - Production/Stable",
    # "Development Status :: 6 - Mature",
    "License :: OSI Approved :: ISC License (ISCL)",
    "Programming Language :: Python :: 3",
]

[project.scripts]
yara-yaml = "yarayaml.cli:main"

[project.urls]
"Bug Tracker" = "https://github.com/dspruell/yara-yaml/issues"
"Homepage" = "https://github.com/dspruell/yara-yaml"
